Fix bugs and add error reporting to analyze subcommands
- Fix integer division in BinCounter.getPosRatio() (long/long → double)
- Fix broken NaN check in ConservationCloudFeature (== → Double.isNaN)
- Fix wrong variable in apo_protein error message (proteinFile → apoProteinFile)
- Fix outerLater typo → outerLayer in Atoms.SphereLayers and usages
- Fix xenegy_cloud2_layered typo → xenergy_cloud2_layered in Params and usages
- Add error reporting (writeErrorCsvs) to all analyze subcommands
- Add ignoreLigandsSwitch to doCmdFasta (doesn't need ligands)

+++ src/main/groovy/cz/siret/prank/utils/BinCounter.groovy
@@ -28,7 +28,7 @@ class BinCounter<K> {
         double getPosRatio() {
             if (count == 0) return 0
 
-            return positives / count
+            return (double) positives / count
         }
     }
